Split the file into more manageable and visible pieces.
